feat(FOROME-1310): rename card repository
