+++ src/pages/main/components/selected-dataset/build-flow/components/build-flow-header.tsx
@@ -8,7 +8,7 @@ import { t } from '@i18n'
 import { datasetStore } from '@store/dataset'
 import dirinfoStore from '@store/dirinfo'
 import { Routes } from '@router/routes.enum'
-import { CardTitle } from '@ui/card-main'
+import { CardTitle } from '@ui/card'
 import { Icon } from '@ui/icon'
 import { DatasetCard } from '@data-testid'
 import { isDev } from '@pages/main/main.constants'
