+++ src/pages/main/components/selected-dataset/build-flow/components/cards/description-card.tsx
@@ -5,7 +5,7 @@ import { ExploreGenomeTypes } from '@core/enum/explore-genome-types-enum'
 import { datasetStore } from '@store/dataset'
 import filterStore from '@store/filter'
 import { Button } from '@ui/button'
-import { Card } from '@ui/card-main'
+import { Card } from '@ui/card'
 import { GlbPagesNames } from '@glb/glb-names'
 import {
   exploreCandidateOptionsList,
