+++ src/pages/main/components/selected-dataset/build-flow/components/cards/presets-card/presets-card.tsx
@@ -6,7 +6,7 @@ import { observer } from 'mobx-react-lite'
 import datasetStore from '@store/dataset/dataset'
 import { Routes } from '@router/routes.enum'
 import { Button } from '@ui/button'
-import { Card, CardTitle } from '@ui/card-main'
+import { Card, CardTitle } from '@ui/card'
 import { Icon } from '@ui/icon'
 import { Loader } from '@ui/loader'
 import { ICardProps } from '../../wizard/wizard.interface'
