+++ src/pages/main/components/selected-dataset/build-flow/components/cards/start-card.tsx
@@ -1,6 +1,6 @@
 import { datasetStore } from '@store/dataset'
 import { Button } from '@ui/button'
-import { Card } from '@ui/card-main'
+import { Card } from '@ui/card'
 import { Radio } from '@ui/radio'
 import { startFlowOptionsList } from '../wizard/wizard.data'
 import { ICardProps } from '../wizard/wizard.interface'
