+++ src/pages/main/components/selected-dataset/components/dataset-info/dataset-info.tsx
@@ -8,7 +8,7 @@ import { observer } from 'mobx-react-lite'
 import { useFullScreenView } from '@core/hooks/use-fullscreen-view'
 import { t } from '@i18n'
 import datasetStore from '@store/dataset/dataset'
-import { Card, CardTitle } from '@ui/card-main'
+import { Card, CardTitle } from '@ui/card'
 import { Loader } from '@ui/loader'
 import { IDsInfo, Versions } from '@service-providers/dataset-level'
 import { INFO } from './dataset-info.constants'
