+++ src/pages/main/components/selected-dataset/start-flow/components/card-sections/card-previous-explore-section.tsx
@@ -3,7 +3,7 @@ import { useHistory } from 'react-router'
 
 import { LocalStoreManager } from '@core/storage-management'
 import { t } from '@i18n'
-import { CardTitle } from '@ui/card-main'
+import { CardTitle } from '@ui/card'
 import { PREVIOUS_WORK_LOCATION } from '../../../build-flow/components/wizard/wizard.utils'
 import { parseLocation } from './card-sections.utils'
