+++ src/pages/main/components/selected-dataset/start-flow/components/card-sections/card-start-explore-section.tsx
@@ -4,7 +4,7 @@ import { observer } from 'mobx-react-lite'
 import { ExploreTypes } from '@core/enum/explore-types-enum'
 import { datasetStore } from '@store/dataset'
 import { Button } from '@ui/button'
-import { CardTitle } from '@ui/card-main'
+import { CardTitle } from '@ui/card'
 import { Radio } from '@ui/radio'
 import { wizardScenarios } from '../../../build-flow/components/wizard/scenarios/wizard-scenarious'
 import wizardStore from '../../../build-flow/components/wizard/wizard.store'
